feat: apply concept of "game world effect"
Specifically for the heavy ball powerup. Most of the other powerups affect the bounce strategy of the paddle. But this one affects something else in world e.g. the bricks change to triggers so the ball powers through instead of bounce. Both kinds of effects will now overwrite each other.